app: Move non-main stuff from emulator to new app library
