+++ src/emulator/app/include/app/screen_render.h
