+++ src/emulator/app/src/app.cpp
@@ -15,9 +15,9 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#include "app.h"
-
 #include "sfo.h"
+
+#include <app/app.h>
 
 #include <gui/functions.h>
 #include <gui/imgui_impl_sdl_gl3.h>
