+++ src/emulator/app/src/init.cpp
@@ -15,11 +15,12 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#include "app.h"
 #include "sfo.h"
 
+#include <app/app.h>
+#include <app/config.h>
+
 #include <audio/functions.h>
-#include <config.h>
 #include <glutil/gl.h>
 #include <host/state.h>
 #include <host/version.h>
